requirements.txt
python-telegram-bot==13.15
fastapi
uvicorn
gspread
oauth2client
